scripts: Add scripts to generate Ghidra parser profile
The current efiSeek has blob Ghidra data type libraries which cannot
reproduce because of lack of source header files and Ghidra parser
profile. The blob gdt files seem to be generated from behemoth.h
included in ida-efiutils [1]. This commit adds helper scripts that
fetch the stable EDK2 source code and generate Ghidra parser profile for
the source code. The script is based on the gen_prf.sh included in
ghidra-firmware-utils [2]. Note that the latest release version of
Ghidra fails to parse vanilla EDK2 source code due to lack of
_Static_assert macro support as reported in issue [3]. This commit also
includes a workaround patch for the issue.

[1] https://github.com/snare/ida-efiutils
[2] https://github.com/al3xtjames/ghidra-firmware-utils
[3] https://github.com/NationalSecurityAgency/ghidra/issues/4038

Signed-off-by: Akira Moroo <retrage01@gmail.com>